Created Compiled Harp Area.
The reason I include the compiled Harp directory is becuase it saves me
the trouble of installing nodejs and a bunch of other junk on my server.
Instead, my server only needs to know how to run C and serve HTML
content for the most part. This is easier, though maybe given some time
I'll write up a make file that will both compile and create the static
website and cgi scripts.