Working on application configuration.
Created urls header to keep track of URL's so that they can be compiled
in directly where applicable.

Made login form point to cgi file, discovered weird issue with my
Makefile where it compiles the heartbeat cgi object for everything. Need
to determine why that is the case and fix it.

Updated sample config with session timeout and base url.