Work on create line items form.
Removed Categories, it doesn't make sense to have them right now. I will
just use seperate 'accounts' to keep track of what spending is going
where. As this is not a banking application per se, but a budgetting
one. And therefore it makes sense to have multiple trackers... which is
more or less what an 'account' is right now. It's a tracking bucket for
line items.

Need to put in geo loc via the navigator.