Updated line items list. Created test directory.
Need to be able to test the various pieces of the scripting without
having a session or other things to be in the way. So a test directory
has been created to facilitate. Right now there's something wrong with
the line item listing as I am going off into bad memory or something
like that.