Added n2c function and cdn for leaflet.
For map page, updated layout with leaflet cdn links, also created small
and simple function to convert an account name into a color to be used.
Right now I dont have many colors in place in the list, but really, it's
pretty simple to extend and I'm not concerned yet.