TroubleShooting Update Account Balance.
Problem with writing out to the accounts file is the issue. The update
itself works fine (adding the number to the specified account and
writing it all out to a temporary file)

But copying the data from the temporary file to the new accounts file is
failing for some reason. Not sure why yet, it doesn't seem like a
streaming problem where there needs to be a seek operation between
input/output to the file because I am calling fflush and also have tried
fseek (and rewind) as well.