Added Sanitize function for filenames.
Replaces ' ' with '-', this works well to prevent spaces in the
filename, however now I am running into the problem of whitespace at the
end of the file names from the account reading process (read with
"%d 64[^0-9] %lf") which is adding an extra space onto the end of the
word, this translates into a -, which is problematic for the filenames
to say the least.

Wrote test to make sure things work as well.