Created Timeline CGI script.
While I can attempt to loop ajax calls to retrieve a bunch of accounts
and lineitems, it really makes it frustrating (in some regards) to have
such convoluted javascript code.

So rather than that, it's much simpler to generate the data I need all
at once and use that. So timeline.c does that pretty easily.