Refactored Timeline.js to use timeline.cgi
It's still doing a month at a time view. I'd really like to do a day by
day view, and I could easily do it for mon-sun, but I'd like to be able
to do it with each day of each month _that are used_ and any inbetween.
Haven't quite figured out the best way to do this yet unfortunately.