Added in day counts graph.
I am thinking I should create a third graph, which will be a bit more
dynamic, I could have a dropdown to choose a month, and then that month
would switch the graph data out for information from that month. It'd be
easy to show a line graph of the spending habits then.