Started on Comparison functionality.
Created HTML template for comparing weeks/months to each other. Simple
pie chart and list of items in each is the plan. With the line items
able to drop down or be hidden.

When clicking to show items in one area, it should drop down the other
ones near it maybe. Not sure yet if I want that as it could be annoying
to have things automatically move on you.

Need to create JS to create pie charts/populate items. I think I will
populate the lists first and then use the data in each of the line items
to compute it. I think that will work well.