Added skeleton for by week function.
Vague plan to create the week by week view:

Data is naturally ordered in first ot last ordering due to the nature
of the storage mechanism. This means that if we read it backwards into
the templates then we will have a natural order. So therefore it is only
neccesary to choose a boundary for the week to fall on. Given that we
live in america... this could be a monday or sunday. Given that I am not
a business man. it should be sunday. But. given that we might visualize
the data in a timeline. we'd want sunday + saturday together. so monday
it will be.