Created Manage Lineitem page
Created the page, not the js yet that will allow for the deletion of
account line items / creation of the table holding them, etc, but yeah,
the page is there at least.

Need to figure out how I'm going to paginate the line items and such.
Need to create an api controller to spit out json for the accounts
themselves and probbaly limit them a bit with offsets, paginations, and
etc etc