Able to create goals
Yay! Now that the goals are done we can get to work on the metrics side
of things, aggregating expendature information and then showing the
goals and how everything stacks up!

Well, after we create the goals management page of course so we can
delete and update the accounts and stuff.