Added Migration Script.
Migration script should be moved into the data directory for the
specific user. That means if there are 3 users, the data directory looks
like this:

data/
 - user1/
   - accounts
   - account1
   - account2 ...
 - user2/
 - user3/

You should move the migration file into the user folder, create the user
in the new bgi, grab their user id, and then set the define in the
script and run it.