Adding note to self.
Note to self.
Check out this gist, the scale.color definitely seems like something I
want to use that will be better than my previous method. The pie chart
isn't SUPER pretty, but that's ok. It's just a visual it doesn't need to
be the most amazing thing since sliced bread.

Need to figure out how we're going to represent the data and goals.

Bar charts make sense somewhat, but when a goal is linked to multiple
accounst then we need an aggregate of both of those accounts to compare
it to. It seems like we'll have a lot of custom charts that way.

If a goal is only linked to one account, then we can simple draw a
horizontal line against the actual amount spent for that account, that
should show a clear visual of how close we are to failing and whatnot.

And of course we can just do it non-visually (besides some red text
perhaps or green) and say you've spent X/Y for goal Q.

The other thing to note as a possible bit is to save the goals and be
able to peruse them to see how we're doing. Consider it your gold star
on life I guess, it would be interesting to catalog if goals are met
each week/month/time period set and see what causes us to fail or meet
goals. Sort of a behavioral analysis in that way, it'd be fun to expand
the whole project into a life tracker of sorts and keep track of my mood
during the day and other things, essentially rating my stresses and
noting each. It'd be fascinating to see if my spending is based around
my moods at all or if I have different outlets (spending more money on
alcohol or sweets or something)