Setting up weekly data for year
Realized I was doing the wrong query. Need to work on the SQL to get
each range of weeks back in each row. Might work to make a week table or
a temporary table like that and somehow do some grouping against it on a
join or something. Will need to research a bitmore.