Add CrudDAO trait to be used by models
The cool thing is that the scaladoc is inherited to the implementing
DAOs. Pretty neat.

Basic DAO to mix into model DAOs so that it's easy to understand basic
DAO operations. Uses Futures to stay async so we can use it with mysql,
mongo, a web service, or whatever we'd like.