PageData model for Dashboard
I think this is right, not sure yet though. It might be better to focus
on the pie models again and pass a list of charts to plot or something
by the view. Ideally I should only have to give things to render to the
view, not do logic for that inside.