Update Schema, Parent child not M2M
Don't know what I was thinking making a many to many, I guess the
concept of tagging got at my head or something. For categories it should
be parent child. Fixed the DAO and schema to reflect this change.

Update DashboardPageData to include charts model. Need to make something
like a chart service or something that performs good database queries to
aggregate and get information for charts.